@@ -11,11 +11,10 @@
     }
   ],
   "require": {
-    "php": ">=5.4.0"
+    "php": ">=7.1"
   },
   "require-dev": {
-    "phpunit/phpunit": "~4.8",
-    "squizlabs/php_codesniffer": "~2.3"
+    "phpunit/phpunit": "^7.0 || ^8.0 || ^10.0"
   },
   "autoload": {
     "psr-4": {
