@@ -14,16 +14,16 @@ use CHItA\PHPDiff\Diff3Algorithm\WeaveMerge;
 use CHItA\PHPDiff\Differ;
 use CHItA\PHPDiff\Differ3;
 use CHItA\PHPDiff\LongestCommonSubsequence\Algorithm\Hirschberg;
-use PHPUnit_Framework_TestCase;
+use PHPUnit\Framework\TestCase;
 
-class DifferBaseTest extends PHPUnit_Framework_TestCase
+class DifferBaseTest extends TestCase
 {
     /**
      * @var Differ3
      */
     private $differ;
 
-    public function setUp()
+    public function setUp(): void
     {
         $this->differ = new Differ3(
             new WeaveMerge(new Differ(null, new Hirschberg()))
